Conform SGS2 SD cards to AOSP standard (1/3)
Same patch as on SGS1, swap their mountpoints but force switchexternal.
Useful for new users coming to CM from stock as this is expected behavior,
but will also fix the logic in CMParts.

Patchset 2: Totally messed up the partition scheme. Anyone who +1'd this before this patch is a liar xD

Change-Id: I7be24dfe0fb65c948781f90375faeec33cf593f8